backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
sqlite3-to-mysql
alembic
pydantic
pydantic-settings
httpx
authlib
itsdangerous
python-multipart
jinja2
python-jose[cryptography]
passlib[bcrypt]